requirements.txt
# MetaProxy-Trader Dependencies
# Автоматизация Android приложений
appium-python-client==4.1.0

# Веб-драйвер для автоматизации
selenium==4.15.2

# MITM прокси для перехвата трафика
mitmproxy==10.1.5

# Дополнительные зависимости для работы с Appium
urllib3==2.0.7
certifi==2023.7.22

# Для работы с HTTP запросами (может потребоваться)
requests==2.31.0

# Для обработки JSON (уже включен в Python, но для совместимости)
# json - встроенная библиотека Python

# Встроенные библиотеки Python (не требуют установки):
# - time
# - random  
# - string
# - os